# source/docs/swagger.yaml
basePath: /
definitions:
  controller.RequestLogin:
    properties:
      email:
        example: admin@bumame.com
        type: string
      password:
        example: "123456"
        type: string
    required:
    - email
    - password
    type: object
  web.ErrorResponse:
    properties:
      messages:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  web.RequestB2BDokter:
    properties:
      dokter_name:
        type: string
    required:
    - dokter_name
    type: object
...
  web.ResponseStandard:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/web.ErrorResponse'
        type: array
      fulfilled:
        type: integer
      message:
        type: string
      pagination:
        description: |-
          Meta       any `json:"meta"`
          JsonApi    any `json:"jsonapi"`
          Links      any `json:"links"`
          Included   any `json:"included"`
          Self       any `json:"self"`
          Related    any `json:"related"`
      statusCode:
        type: integer
    type: object
host: localhost:8200
info:
  contact: {}
  description: This is a sample server with v1 and v2 endpoints.
  title: Sarana Boilerplate API
  version: "2.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.RequestLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Login
      tags:
      - auth
  /auth/read-token:
    get:
      consumes:
      - application/json
      description: Read Token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Read Token
      tags:
      - auth
  /b2b-dokter:
    get:
      consumes:
      - application/json
      description: Get all b2b dokter
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Sort
        in: query
        name: sort
        type: string
      - in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Get all b2b dokter
      tags:
      - b2b-dokter
    post:
      consumes:
      - application/json
      description: Create b2b dokter
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web.RequestB2BDokter'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Create b2b dokter
      tags:
      - b2b-dokter
  /b2b-dokter/{b2b_dokter_id}:
    delete:
      consumes:
      - application/json
      description: Delete b2b dokter
      parameters:
      - description: B2B dokter ID
        in: path
        name: b2b_dokter_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Delete b2b dokter
      tags:
      - b2b-dokter
    get:
      consumes:
      - application/json
      description: Get b2b dokter by ID
      parameters:
      - description: B2B dokter ID
        in: path
        name: b2b_dokter_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Get b2b dokter by ID
      tags:
      - b2b-dokter
    put:
      consumes:
      - application/json
      description: Update b2b dokter
      parameters:
      - description: B2B dokter ID
        in: path
        name: b2b_dokter_id
        required: true
        type: string
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web.RequestB2BDokter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Update b2b dokter
      tags:
      - b2b-dokter
  /b2b-dokter/name/{b2b_dokter_name}:
    get:
      consumes:
      - application/json
      description: Get b2b dokter by name
      parameters:
      - description: B2B dokter name
        in: path
        name: b2b_dokter_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ResponseStandard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ResponseStandard'
      summary: Get b2b dokter by name
      tags:
      - b2b-dokter
...
swagger: "2.0"